refactor(core)!: delete legacy auth machinery (AuthManager, CrossDomainAuth, FedCM sign-in, sso/silent/redirect, oxy_rt client) — core 6.0.0
Both SDK cutovers (#528 auth-sdk, #529 services) are merged + deployed, so the
legacy core auth modules now have ZERO SDK callers. Verified each deletion with
a workspace-wide `git grep` before removing.

DELETED (zero live callers):
- AuthManager.ts + AuthManagerTypes.ts + tests (AuthManagerAccount → auth-sdk's
  DeviceAccountView; no core-type importers remain).
- CrossDomainAuth.ts + test.
- mixins/OxyServices.silent.ts, OxyServices.sso.ts, OxyServices.redirect.ts + tests.
- FedCM SIGN-IN surface (OxyServices.fedcm.ts) — signInWithFedCM /
  silentSignInWithFedCM / requestIdentityCredential / isFedCMSupported /
  revokeFedCMCredential / exchangeIdTokenForSession / nonce mint, etc.
- utils/ssoReturn.ts, ssoEstablish.ts + tests.
- base.establishDeviceRefreshSlot (+ the constructor authWebUrl auto-detect).
- auth.ts refresh-cookie client methods: refreshAllSessions,
  refreshTokenViaCookie/_refreshCookieRaw, logoutSessionByAuthuser,
  logoutAllSessionsViaCookie + RefreshAllOptions.
- RefreshAll* CLIENT types (models/interfaces) + accountUtils
  mergeAccountsFromRefreshAll (dead; contracts keeps its own refreshAll schemas).

KEPT (moved to a small OxyServices.authorizedApps mixin) — "Connected apps"
management (listAuthorizedApps/revokeAuthorizedApp + AuthorizedApp): still live
in services' ConnectedAppsScreen until the AppGrant migration.

TRIMMED, NOT deleted (server/IdP still import these — lista B, gated on the
ecosystem bump; deleting would break the api + IdP builds):
- fapiAutoDetect.ts → kept registrableApex (api deviceAuth/sso.controller/
  fedcm.service + IdP + @oxyhq/core/server); dropped autoDetectAuthWebUrl.
- authWebUrl.ts → kept CENTRAL_IDP_APEX/CENTRAL_AUTH_URL (IdP); dropped
  resolveCentralAuthUrl.
- ssoBounce.ts → kept SSO_CALLBACK_PATH (api/IdP via core/server); dropped every
  client bounce helper + getSsoCallbackBootstrapScript.

Cross-package (obsolete call sites; lead-authorised):
- auth-sdk useCommonsSignIn: removed the establishIdpSessionAfterClaim hop
  (device-flow claims now persist rotating refresh tokens) + updated its test.
- accounts/commons/inbox +html.tsx: removed the getSsoCallbackBootstrapScript
  injection (the callback path is dead).
- core accounts-mixin switch: removed the establishDeviceRefreshSlot call (switch
  registers server-side now).
- isWebBrowser: moved to @oxyhq/core (utils/platform), services + auth-sdk local
  copies now re-export core's.

Version pre-stage for F5 publish: core 6.0.0, auth-sdk 8.0.0 (peer/dep ^6.0.0),
services 15.0.0 (dep ^6.0.0); lockfile in this commit.

Gates: core 673 + auth-sdk 71 + services 185 tests green; core (dual+ESM,
require()-free) / auth-sdk (dual) / services (bob) builds green; console vite
build green; no consumer references any deleted symbol.
