+++ packages/auth-sdk/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@oxyhq/auth",
-  "version": "7.0.0",
+  "version": "8.0.0",
   "description": "OxyHQ Web Authentication SDK — headless auth with React hooks for Next.js, Vite, and web apps",
   "main": "dist/cjs/index.js",
   "module": "dist/esm/index.js",
@@ -79,7 +79,7 @@
     "qrcode": "^1.5.4"
   },
   "peerDependencies": {
-    "@oxyhq/core": "^5.0.0",
+    "@oxyhq/core": "^6.0.0",
     "@tanstack/query-sync-storage-persister": "^5.100",
     "@tanstack/react-query": "^5.100",
     "@tanstack/react-query-persist-client": "^5.100",
