+++ packages/core/src/mixins/OxyServices.auth.ts
@@ -3,12 +3,7 @@
  *
  * Supports password-based login (email/username) and public key challenge-response.
  */
-import type {
-  User,
-  RefreshAllResponse,
-  RefreshAllAccount,
-  RefreshCookieResponse,
-} from '../models/interfaces';
+import type { User } from '../models/interfaces';
 import type { UserNameResponse, LoginResult, LoginSessionResult } from '@oxyhq/contracts';
 import { loginResultSchema, safeParseContract } from '@oxyhq/contracts';
 import type { SessionLoginResponse } from '../models/session';
@@ -31,26 +26,6 @@ const COMMONS_SIGN_IN_EXPIRY_MS = 5 * 60 * 1000;
 export interface ChallengeResponse {
   challenge: string;
   expiresAt: string;
-}
-
-/**
- * Options for {@link refreshAllSessions}.
- */
-export interface RefreshAllOptions {
-  /**
-   * Abort the `POST /auth/refresh-all` request after this many milliseconds and
-   * resolve as "no signed-in accounts" (`{ accounts: [] }`) rather than hanging.
-   *
-   * Why: on a cross-domain RP (e.g. `mention.earth`) the `Domain=oxy.so` refresh
-   * cookie never reaches `api.<apex>`, so this request can stall behind a slow
-   * or unreachable endpoint with no useful answer coming back. As one step of
-   * the ordered cold-boot sequence, a stalled refresh is dead latency in front
-   * of the steps that actually hold the answer. A bounded abort lets cold boot
-   * fall through quickly. Omit (or pass `0`/negative) to wait indefinitely.
-   * The abort is treated identically to a 401 — the
-   * "not signed in on this device" path — never an error.
-   */
-  timeout?: number;
 }
 
 export interface RegistrationRequest {
@@ -936,232 +911,6 @@ export function OxyServicesAuthMixin<T extends typeof OxyServicesBase>(Base: T) 
       } catch (error) {
         throw this.handleError(error);
       }
-    }
-
-    /**
-     * Refresh every device-local refresh-cookie slot in a single round trip
-     * (Google-style multi-account rebuild).
-     *
-     * Calls `POST {sessionBaseUrl}/auth/refresh-all` with `credentials: 'include'`
-     * and NO bearer. The browser attaches every indexed `oxy_rt_${authuser}`
-     * cookie it has; the server rotates each in parallel and returns one entry
-     * per VALID account.
-     *
-     * Failure handling:
-     * - 401 → no signed-in accounts on this device → returns `{ accounts: [] }`
-     *   (NOT an error; this is the cold-boot "not signed in" path).
-     * - 404 → endpoint not deployed → returns `{ accounts: [] }`.
-     * - Any other non-2xx → throws via `handleError`.
-     *
-     * The refresh cookie itself never enters JS — only the rotated access
-     * tokens do. Each access token still needs to be planted via
-     * `setTokens(...)` (or per-account in-memory storage) at the consumer.
-     */
-    async refreshAllSessions(options: RefreshAllOptions = {}): Promise<RefreshAllResponse> {
-      const url = `${this.getSessionBaseUrl().replace(/\/$/, '')}/auth/refresh-all`;
-
-      // Optional bounded abort (see `RefreshAllOptions.timeout`). A positive
-      // timeout arms an `AbortController` that aborts the in-flight request; an
-      // abort is treated as "no signed-in accounts on this device" — the same
-      // outcome as a 401 — so a cross-domain stall falls through cleanly instead
-      // of hanging the cold boot.
-      const timeout = typeof options.timeout === 'number' && options.timeout > 0 ? options.timeout : undefined;
-      const controller = timeout !== undefined ? new AbortController() : undefined;
-      const timeoutId = timeout !== undefined && controller
-        ? setTimeout(() => controller.abort(), timeout)
-        : undefined;
-
-      let response: Response;
-      try {
-        response = await fetch(url, {
-          method: 'POST',
-          credentials: 'include',
-          headers: { Accept: 'application/json' },
-          signal: controller?.signal,
-        });
-      } catch (error) {
-        // A bounded-timeout abort is the "not signed in / cross-domain stall"
-        // path, NOT an error. The browser raises a DOMException named
-        // 'AbortError' (some runtimes use a generic Error); match on the name so
-        // we never throw the timeout into the cold-boot error handler.
-        if (error instanceof Error && error.name === 'AbortError') {
-          return { accounts: [] };
-        }
-        throw this.handleError(error);
-      } finally {
-        if (timeoutId !== undefined) {
-          clearTimeout(timeoutId);
-        }
-      }
-
-      if (response.status === 401) {
-        return { accounts: [] };
-      }
-
-      if (response.status === 404) {
-        return { accounts: [] };
-      }
-
-      if (!response.ok) {
-        throw this.handleError(
-          new Error(`Refresh-all failed with HTTP ${response.status}`)
-        );
-      }
-
-      const payload = (await response.json()) as { accounts?: unknown };
-      const raw = Array.isArray(payload.accounts) ? payload.accounts : [];
-      const accounts: RefreshAllAccount[] = [];
-
-      for (const entry of raw) {
-        if (entry === null || typeof entry !== 'object') {
-          continue;
-        }
-        const e = entry as {
-          authuser?: number;
-          accessToken?: string;
-          expiresAt?: string;
-          sessionId?: string;
-          user?: { id?: string; _id?: string; username?: string; name?: UserNameResponse; avatar?: string | null; email?: string; color?: string | null };
-        };
-        if (!e.accessToken || !e.expiresAt || !e.sessionId || !e.user) {
-          continue;
-        }
-        const userId = e.user.id ?? e.user._id;
-        // `name.displayName` is optional on the contract — do NOT drop no-name
-        // accounts. Only an absent userId or username makes the entry unusable.
-        if (!userId || !e.user.username) {
-          continue;
-        }
-        if (typeof e.authuser !== 'number') {
-          continue;
-        }
-        accounts.push({
-          authuser: e.authuser,
-          accessToken: e.accessToken,
-          expiresAt: e.expiresAt,
-          sessionId: e.sessionId,
-          user: {
-            id: userId,
-            username: e.user.username,
-            // `name.displayName` is optional; carry whatever structured name the
-            // server sent (possibly an empty object) and let consumers fall back.
-            name: e.user.name ?? {},
-            avatar: e.user.avatar ?? null,
-            email: e.user.email,
-            color: e.user.color ?? null,
-          },
-        });
-      }
-
-      return { accounts };
-    }
-
-    /**
-     * Rotate a single refresh-cookie slot and return the fresh access token.
-     *
-     * When `authuser` is provided, the server rotates ONLY that slot
-     * (`oxy_rt_${authuser}`) — sibling accounts on the same device stay
-     * untouched. When omitted, the server picks the lowest indexed slot present.
-     * The refresh cookie itself never enters JS.
-     *
-     * Returns `null` on 401 (no cookie / expired / reused) so the caller can
-     * fall through cleanly to the unauthenticated path.
-     */
-    async refreshTokenViaCookie(
-      opts: { authuser?: number } = {}
-    ): Promise<RefreshCookieResponse | null> {
-      const result = await this._refreshCookieRaw(opts.authuser);
-      return result;
-    }
-
-    /**
-     * Sign out a single device-local account by its authuser slot index.
-     *
-     * Revokes that slot's refresh-token family and deactivates its session;
-     * sibling indexed slots stay signed in. The browser-side `oxy_rt_${n}`
-     * cookie is cleared by the server's `Set-Cookie` response header.
-     */
-    async logoutSessionByAuthuser(authuser: number): Promise<void> {
-      const url = `${this.getSessionBaseUrl().replace(/\/$/, '')}/auth/logout?authuser=${encodeURIComponent(String(authuser))}`;
-      try {
-        const response = await fetch(url, {
-          method: 'POST',
-          credentials: 'include',
-          headers: { Accept: 'application/json' },
-        });
-        if (!response.ok && response.status !== 401) {
-          throw new Error(`Logout (authuser=${authuser}) failed with HTTP ${response.status}`);
-        }
-      } catch (error) {
-        throw this.handleError(error);
-      }
-    }
-
-    /**
-     * Sign out EVERY device-local account on this device by clearing every
-     * presented refresh-cookie slot at once. Revokes every family + clears
-     * every slot. Always succeeds (idempotent on unknown/garbage tokens).
-     */
-    async logoutAllSessionsViaCookie(): Promise<void> {
-      const url = `${this.getSessionBaseUrl().replace(/\/$/, '')}/auth/logout`;
-      try {
-        const response = await fetch(url, {
-          method: 'POST',
-          credentials: 'include',
-          headers: { Accept: 'application/json' },
-        });
-        if (!response.ok && response.status !== 401) {
-          throw new Error(`Logout-all failed with HTTP ${response.status}`);
-        }
-      } catch (error) {
-        throw this.handleError(error);
-      }
-    }
-
-    /**
-     * Internal: raw `POST /auth/refresh[?authuser=N]` call returning the
-     * minted access token. Returns `null` on 401 / non-2xx.
-     *
-     * @internal
-     */
-    async _refreshCookieRaw(authuser?: number): Promise<RefreshCookieResponse | null> {
-      const base = this.getSessionBaseUrl().replace(/\/$/, '');
-      const url = typeof authuser === 'number'
-        ? `${base}/auth/refresh?authuser=${encodeURIComponent(String(authuser))}`
-        : `${base}/auth/refresh`;
-
-      let response: Response;
-      try {
-        response = await fetch(url, {
-          method: 'POST',
-          credentials: 'include',
-          headers: { Accept: 'application/json' },
-        });
-      } catch (error) {
-        throw this.handleError(error);
-      }
-
-      if (!response.ok) {
-        return null;
-      }
-
-      const payload = (await response.json()) as {
-        accessToken?: unknown;
-        expiresAt?: unknown;
-        authuser?: unknown;
-      };
-      if (typeof payload.accessToken !== 'string' || !payload.accessToken) {
-        return null;
-      }
-      const expiresAt = typeof payload.expiresAt === 'string' ? payload.expiresAt : '';
-      if (typeof payload.authuser !== 'number') {
-        return null;
-      }
-      return {
-        accessToken: payload.accessToken,
-        expiresAt,
-        authuser: payload.authuser,
-      };
     }
 
     /**
